Squashed commit of the following:
commit c8445ad9b2e9ab8cab50661090a172275bd84e8a
Author: Andrew Shkrob <andrew.shkrob.social@yandex.by>
Date:   Sun Nov 20 20:41:11 2022 +0100

    Merge branch 'master' into android-auto/new

+++ drape_frontend/read_manager.hpp
@@ -106,7 +106,7 @@ private:
 
   void CancelTileInfo(std::shared_ptr<TileInfo> const & tileToCancel);
   void ClearTileInfo(std::shared_ptr<TileInfo> const & tileToClear);
-  void IncreaseCounter(int value);
+  void IncreaseCounter(size_t value);
   void CheckFinishedTiles(TTileInfoCollection const & requestedTiles, bool forceUpdateUserMarks);
 };
 }  // namespace df
